+++ docs/conf.py
@@ -11,9 +11,9 @@ import sys
 sys.path.insert(0, os.path.abspath("../src"))
 
 project = "gh-sprint-report"
-copyright = "gh-sprint-report contributors"
-author = "gh-sprint-report contributors"
-release = "1.0.0"
+copyright = "R Software & Consulting LLC"
+author = "R Software & Consulting LLC"
+release = "1.0.2"
 
 extensions = [
     "sphinx.ext.autodoc",